requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
black==23.1.0
boto3==1.21.22
botocore==1.24.22
certifi==2021.10.8
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
distlib==0.3.6
Django==3.2.12
django-allauth==0.49.0
django-braces==1.15.0
django-colorfield==0.6.3
django-cors-headers==3.11.0
django-filter==22.1
django-import-export==3.1.0
django-ipware==4.0.2
django-oauth-toolkit==1.7.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-storages==1.12.3
django-tinymce==3.4.0
djangorestframework==3.13.1
drf-spectacular==0.25.1
et-xmlfile==1.1.0
filelock==3.9.0
flake8==6.0.0
frozenlist==1.3.3
identify==2.5.18
idna==3.4
inflection==0.5.1
install==1.3.5
isort==5.12.0
jmespath==1.0.1
jsonschema==4.17.3
jwcrypto==1.4.2
MarkupPy==1.14
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.24.1
oauthlib==3.2.2
odfpy==1.4.1
openai==0.26.5
openpyxl==3.1.1
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
pre-commit==3.1.1
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tablib==3.3.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.20.0
whitenoise==6.3.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
